requirements.txt
Flask==2.3.3
numpy==1.24.3
opencv-python==4.8.0.76
scikit-learn==1.3.0
tensorflow==2.13.0
Pillow==10.0.0
joblib==1.3.2
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2

